Fixed requested changes.
