+++ kairon/shared/data/processor.py
@@ -50,7 +50,7 @@ from kairon.shared.importer.processor import DataImporterLogProcessor
 from kairon.shared.metering.constants import MetricType
 from kairon.shared.metering.metering_processor import MeteringProcessor
 from kairon.shared.models import StoryEventType, TemplateType, StoryStepType, HttpContentType, StoryType, \
-    LlmPromptSource, CognitionDataType
+    LlmPromptSource
 from kairon.shared.plugins.factory import PluginFactory
 from kairon.shared.utils import Utility, StoryValidator
 from .constant import (
@@ -89,8 +89,8 @@ from .data_objects import (
     ModelDeployment,
     Rules,
     Utterances, BotSettings, ChatClientConfig, SlotMapping, KeyVault, EventConfig, TrainingDataGenerator,
-    MultiflowStories, MultiflowStoryEvents, CognitionData, MultiFlowStoryMetadata,
-    Synonyms, Lookup, Analytics, CognitionSchema, ColumnMetadata
+    MultiflowStories, MultiflowStoryEvents, MultiFlowStoryMetadata,
+    Synonyms, Lookup, Analytics
 )
 from .utils import DataUtility
 from ..constants import KaironSystemSlots, PluginTypes
@@ -5757,212 +5757,3 @@ class MongoProcessor:
             action.pop('user')
 
             yield action
-
-    def is_collection_limit_exceeded(self, bot, collection):
-        """
-        checks if collection limit is exhausted
-
-        :param bot: bot id
-        :param collection: Name of collection
-        :return: boolean
-        :raises: AppException
-        """
-
-        collections = self.list_collection(bot)
-        doc_count = CognitionData.objects(
-            bot=bot, collection__ne=None,
-        ).count()
-        if collection not in collections and doc_count >= BotSettings.objects(bot=bot).get().collection_limit:
-            return True
-        else:
-            return False
-
-    def save_content(self, content: Text, user: Text, bot: Text, collection: Text = None):
-        if collection:
-            if self.is_collection_limit_exceeded(bot, collection):
-                raise AppException('Collection limit exceeded!')
-        bot_settings = self.get_bot_settings(bot=bot, user=user)
-        if not bot_settings["llm_settings"]['enable_faq']:
-            raise AppException('Faq feature is disabled for the bot! Please contact support.')
-        if len(content.split()) < 10:
-            raise AppException("Content should contain atleast 10 words.")
-
-        content_obj = CognitionData()
-        content_obj.data = content
-        content_obj.collection = collection
-        content_obj.user = user
-        content_obj.bot = bot
-        id = (
-            content_obj.save().id.__str__()
-        )
-        return id
-
-    def update_content(self, content_id: str, content: Text, user: Text, bot: Text, collection: Text = None):
-        if len(content.split()) < 10:
-            raise AppException("Content should contain atleast 10 words.")
-
-        Utility.is_exist(CognitionData, bot=bot, id__ne=content_id, data=content, content_type__ne=CognitionDataType.json.value,
-                         exp_message="Text already exists!")
-
-        try:
-            content_obj = CognitionData.objects(bot=bot, id=content_id).get()
-            content_obj.data = content
-            content_obj.collection = collection
-            content_obj.user = user
-            content_obj.timestamp = datetime.utcnow()
-            content_obj.save()
-        except DoesNotExist:
-            raise AppException("Content with given id not found!")
-
-    def delete_content(self, content_id: str, user: Text, bot: Text):
-        try:
-            content = CognitionData.objects(bot=bot, id=content_id).get()
-            content.delete()
-        except DoesNotExist:
-            raise AppException("Text does not exists!")
-
-    def get_content(self, bot: Text, **kwargs):
-        """
-        fetches content
-
-        :param bot: bot id
-        :return: yield dict
-        """
-        kwargs["bot"] = bot
-        search = kwargs.pop('data', None)
-        start_idx = kwargs.pop('start_idx', 0)
-        page_size = kwargs.pop('page_size', 10)
-        cognition_data = CognitionData.objects(**kwargs)
-        if search:
-            cognition_data = cognition_data.search_text(search)
-        for value in cognition_data.skip(start_idx).limit(page_size):
-            item = value.to_mongo().to_dict()
-            item.pop('timestamp')
-            item["_id"] = item["_id"].__str__()
-            yield item
-
-    def list_collection(self, bot: Text):
-        """
-        Retrieve cognition data.
-        :param bot: bot id
-        """
-        collections = list(CognitionData.objects(bot=bot).distinct(field='collection'))
-        return collections
-
-    def save_cognition_schema(self, metadata: Dict, user: Text, bot: Text):
-        column_name = [meta['column_name'] for meta in metadata.get('metadata')]
-        for column in column_name:
-            Utility.is_exist(CognitionSchema, bot=bot, metadata__column_name=column,
-                             exp_message="Column already exists!")
-        metadata_obj = CognitionSchema(bot=bot, user=user)
-        metadata_obj.metadata = [ColumnMetadata(**meta) for meta in metadata.get('metadata')]
-        metadata_obj.collection_name = metadata.get('collection_name', None)
-        metadata_id = metadata_obj.save().to_mongo().to_dict()["_id"].__str__()
-        return metadata_id
-
-    def update_cognition_schema(self, metadata_id: str, metadata: Dict, user: Text, bot: Text):
-        metadata_items = metadata.get('metadata')
-        Utility.is_exist(CognitionSchema, bot=bot, id__ne=metadata_id, metadata=metadata_items,
-                         exp_message="Schema already exists!")
-        try:
-            metadata_obj = CognitionSchema.objects(bot=bot, id=metadata_id).get()
-            metadata_obj.metadata = [ColumnMetadata(**meta) for meta in metadata.get('metadata')]
-            metadata_obj.collection_name = metadata.get('collection_name')
-            metadata_obj.bot = bot
-            metadata_obj.user = user
-            metadata_obj.timestamp = datetime.utcnow()
-            metadata_obj.save()
-        except DoesNotExist:
-            raise AppException("Schema with given id not found!")
-
-    def delete_cognition_schema(self, metadata_id: str, bot: Text):
-        try:
-            metadata = CognitionSchema.objects(bot=bot, id=metadata_id).get()
-            metadata.delete()
-        except DoesNotExist:
-            raise AppException("Schema does not exists!")
-
-    def list_cognition_schema(self, bot: Text):
-        """
-        fetches metadata
-
-        :param bot: bot id
-        :return: yield dict
-        """
-        for value in CognitionSchema.objects(bot=bot):
-            final_data = {}
-            item = value.to_mongo().to_dict()
-            metadata = item.pop("metadata")
-            collection = item.pop('collection_name', None)
-            final_data["_id"] = item["_id"].__str__()
-            final_data['metadata'] = metadata
-            final_data['collection_name'] = collection
-            yield final_data
-
-    def __validate_metadata_and_payload(self, payload, bot: Text):
-        metadata = list(self.list_cognition_schema(bot))
-        data = payload.get('data')
-        collection = payload.get('collection', None)
-        matched_metadata = Utility.find_matching_metadata(data, metadata, collection)
-        if not matched_metadata:
-            raise AppException("Metadata related to payload not found!")
-        results = [metadata_entry for metadata_dict in matched_metadata for metadata_entry in metadata_dict['metadata']]
-        for metadata_dict in results:
-            Utility.retrieve_data(data, metadata_dict)
-
-    def save_cognition_data(self, payload: Dict, user: Text, bot: Text):
-        bot_settings = self.get_bot_settings(bot=bot, user=user)
-        if not bot_settings["llm_settings"]['enable_faq']:
-            raise AppException('Faq feature is disabled for the bot! Please contact support.')
-        if payload.get('content_type') != CognitionDataType.text.value:
-            self.__validate_metadata_and_payload(payload, bot)
-        payload_obj = CognitionData()
-        payload_obj.data = payload.get('data')
-        payload_obj.content_type = payload.get('content_type')
-        payload_obj.collection = payload.get('collection', None)
-        payload_obj.user = user
-        payload_obj.bot = bot
-        payload_id = payload_obj.save().to_mongo().to_dict()["_id"].__str__()
-        return payload_id
-
-    def update_cognition_data(self, payload_id: str, payload: Dict, user: Text, bot: Text):
-        data = payload['data']
-        content_type = payload['content_type']
-        Utility.is_exist(CognitionData, bot=bot, id__ne=payload_id, data=data, content_type__ne=CognitionDataType.json.value,
-                         exp_message="Payload data already exists!")
-
-        try:
-            payload_obj = CognitionData.objects(bot=bot, id=payload_id).get()
-            payload_obj.data = data
-            payload_obj.content_type = content_type
-            payload_obj.collection = payload.get('collection', None)
-            payload_obj.user = user
-            payload_obj.timestamp = datetime.utcnow()
-            payload_obj.save()
-        except DoesNotExist:
-            raise AppException("Payload with given id not found!")
-
-    def delete_cognition_data(self, payload_id: str, bot: Text):
-        try:
-            payload = CognitionData.objects(bot=bot, id=payload_id).get()
-            payload.delete()
-        except DoesNotExist:
-            raise AppException("Payload does not exists!")
-
-    def list_cognition_data(self, bot: Text):
-        """
-        fetches content
-
-        :param bot: bot id
-        :return: yield dict
-        """
-        for value in CognitionData.objects(bot=bot):
-            final_data = {}
-            item = value.to_mongo().to_dict()
-            data = item.pop("data")
-            data_type = item.pop("content_type")
-            final_data["_id"] = item["_id"].__str__()
-            final_data['content'] = data
-            final_data['content_type'] = data_type
-            final_data['collection'] = item.get('collection', None)
-            yield final_data
